refactor(chrome): move item match logic to ChromeUtils
Extracted item match scoring functions to ChromeUtils for reuse.
Added WallpaperAppId to ChromeSettings and supporting extension methods.
Updated DefaultSidebarChrome to use centralized match logic and support wallpaper app.
