+++ Ivy/Chrome/DefaultSidebarChrome.cs
@@ -8,7 +8,6 @@ using Ivy.Shared;
 using Ivy.Views;
 using Ivy.Widgets.Internal;
 using System.Collections.Immutable;
-using System.Linq;
 
 namespace Ivy.Chrome;
 
@@ -18,40 +17,6 @@ public class DefaultSidebarChrome(ChromeSettings settings) : ViewBase
     private record TabState(string Id, string AppId, string Title, AppHost AppHost, Icons? Icon, string RefreshToken)
     {
         public Tab ToTab() => new Tab(Title, AppHost).Icon(Icon).Key(Utils.GetShortHash(Id + RefreshToken));
-    }
-    private static bool IsWordMatch(string tag, string searchString)
-    {
-        // Split the tag into individual words (separated by hyphens, underscores, or spaces)
-        var words = System.Text.RegularExpressions.Regex.Split(tag, @"[-_\s]+");
-
-        // Check if any word starts with the search string (prefix matching)
-        return words.Any(word => word.StartsWith(searchString, StringComparison.OrdinalIgnoreCase));
-    }
-
-    private int itemMatchScore(MenuItem item, string searchString)
-    {
-        var label = item.Label ?? "";
-
-        // Exact match gets highest priority (score 3)
-        if (string.Equals(label, searchString, StringComparison.OrdinalIgnoreCase))
-        {
-            return 3;
-        }
-
-        // Label contains search string gets medium priority (score 2)
-        if (label.Contains(searchString, StringComparison.OrdinalIgnoreCase))
-        {
-            return 2;
-        }
-
-        // Search hints match gets lowest priority (score 1)
-        if (item.SearchHints?.Any(tag => IsWordMatch(tag, searchString)) == true)
-        {
-            return 1;
-        }
-
-        // No match
-        return 0;
     }
 
     public override object? Build()
@@ -87,7 +52,7 @@ public class DefaultSidebarChrome(ChromeSettings settings) : ViewBase
             {
                 var result = appRepository.GetMenuItems().Flatten()
                     .Where(item => item.Children == null || item.Children.Length == 0) // Only include leaf nodes (actual apps)
-                    .Select(item => new { Item = item, Score = itemMatchScore(item, search.Value) })
+                    .Select(item => new { Item = item, Score = ChromeUtils.ItemMatchScore(item, search.Value) })
                     .Where(x => x.Score > 0)
                     .OrderByDescending(x => x.Score)
                     .ThenBy(x => x.Item.Label)
@@ -342,9 +307,20 @@ public class DefaultSidebarChrome(ChromeSettings settings) : ViewBase
         }
         else
         {
-            body = new TabsLayout(OnTabSelect, OnTabClose, OnTabRefresh, OnTabReorder, selectedIndex.Value,
-                tabs.Value.ToArray().Select(e => e.ToTab()).ToArray()
-            ).RemoveParentPadding().Variant(TabsVariant.Tabs).Padding(0);
+            if (tabs.Value.Length == 0)
+            {
+                body = null;
+                if (settings.WallpaperAppId != null)
+                {
+                    body = new AppHost(settings.WallpaperAppId, null, args.ConnectionId);
+                }
+            }
+            else
+            {
+                body = new TabsLayout(OnTabSelect, OnTabClose, OnTabRefresh, OnTabReorder, selectedIndex.Value,
+                    tabs.Value.ToArray().Select(e => e.ToTab()).ToArray()
+                ).RemoveParentPadding().Variant(TabsVariant.Tabs).Padding(0);
+            }
         }
 
         var searchInput = search.ToSearchInput().ShortcutKey("CTRL+K").TestId("sidebar-search");
