+++ Ivy/Chrome/Shared.cs
@@ -5,6 +5,7 @@ using Ivy.Client;
 using Ivy.Core;
 using Ivy.Core.Hooks;
 using Ivy.Hooks;
+using Ivy.Shared;
 
 namespace Ivy.Chrome;
 
@@ -14,11 +15,48 @@ public enum ChromeNavigation
     Pages
 }
 
+public static class ChromeUtils
+{
+    private static bool IsWordMatch(string tag, string searchString)
+    {
+        var words = System.Text.RegularExpressions.Regex.Split(tag, @"[-_\s]+");
+        return words.Any(word => word.StartsWith(searchString, StringComparison.OrdinalIgnoreCase));
+    }
+
+    public static int ItemMatchScore(MenuItem item, string searchString)
+    {
+        var label = item.Label ?? "";
+
+        // Exact match gets highest priority (score 3)
+        if (string.Equals(label, searchString, StringComparison.OrdinalIgnoreCase))
+        {
+            return 3;
+        }
+
+        // Label contains search string gets medium priority (score 2)
+        if (label.Contains(searchString, StringComparison.OrdinalIgnoreCase))
+        {
+            return 2;
+        }
+
+        // Search hints match gets lowest priority (score 1)
+        if (item.SearchHints?.Any(tag => IsWordMatch(tag, searchString)) == true)
+        {
+            return 1;
+        }
+
+        // No match
+        return 0;
+    }
+}
+
+
 public record ChromeSettings
 {
     public object? Header { get; init; }
     public object? Footer { get; init; }
     public string? DefaultAppId { get; init; }
+    public string? WallpaperAppId { get; init; }
     public bool PreventTabDuplicates { get; init; }
     public ChromeNavigation Navigation { get; init; }
 
@@ -39,6 +77,13 @@ public static class ChromeSettingsExtensions
         var type = typeof(T);
         var descriptor = AppHelpers.GetApp(type);
         return settings with { DefaultAppId = descriptor.Id };
+    }
+    public static ChromeSettings WallpaperAppId(this ChromeSettings settings, string? wallpaperAppId) => settings with { WallpaperAppId = wallpaperAppId };
+    public static ChromeSettings WallpaperApp<T>(this ChromeSettings settings)
+    {
+        var type = typeof(T);
+        var descriptor = AppHelpers.GetApp(type);
+        return settings with { WallpaperAppId = descriptor.Id };
     }
     public static ChromeSettings Navigation(this ChromeSettings settings, ChromeNavigation navigation) => settings with { Navigation = navigation };
     public static ChromeSettings UseTabs(this ChromeSettings settings, bool preventDuplicates = false) => settings with { Navigation = ChromeNavigation.Tabs, PreventTabDuplicates = preventDuplicates };
